pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shareclaw"
version = "0.1.0"
description = "Shared WeChat seat management for cloud-hosted OpenClaw AI assistants"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "garden"},
]
keywords = ["openclaw", "wechat", "lighthouse", "tencent-cloud", "customer-service"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
]
dependencies = [
    "tencentcloud-sdk-python>=3.0.0",
    "flask>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "build>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/gardennchen/ShareClaw"
Repository = "https://github.com/gardennchen/ShareClaw"
Issues = "https://github.com/gardennchen/ShareClaw/issues"

[project.scripts]
shareclaw = "shareclaw.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["shareclaw"]
